+++ samila/functions.py
@@ -6,7 +6,7 @@ import io
 import json
 import random
 import matplotlib
-from .params import DEFAULT_START, DEFAULT_STOP, DEFAULT_STEP, DEFAULT_COLOR, DEFAULT_IMAGE_SIZE
+from .params import DEFAULT_START, DEFAULT_STOP, DEFAULT_STEP, DEFAULT_COLOR, DEFAULT_IMAGE_SIZE, DEFAULT_DEPTH
 from .params import DEFAULT_BACKGROUND_COLOR, DEFAULT_SPOT_SIZE, DEFAULT_PROJECTION, DEFAULT_ALPHA, DEFAULT_LINEWIDTH
 from .params import Projection, VALID_COLORS, NFT_STORAGE_API, OVERVIEW
 from .params import DATA_TYPE_ERROR, CONFIG_TYPE_ERROR, PLOT_DATA_ERROR, CONFIG_NO_STR_FUNCTION_ERROR
@@ -233,6 +233,19 @@ def generate_params_filter(
     g.seed, g.start, g.step, g.stop = seed, start, step, stop
 
 
+def save_params_filter(g, depth=None):
+    """
+    Filter save_image method parameters.
+
+    :param depth: depth of image
+    :type depth: float
+    :return: None
+    """
+    if depth is None:
+        depth = g.depth
+    g.depth = depth
+
+
 def _GI_initializer(g, function1, function2):
     """
     Initialize the generative image.
@@ -264,6 +277,7 @@ def _GI_initializer(g, function1, function2):
     g.projection = DEFAULT_PROJECTION
     g.alpha = DEFAULT_ALPHA
     g.linewidth = DEFAULT_LINEWIDTH
+    g.depth = DEFAULT_DEPTH
 
 
 def nft_storage_upload(api_key, data):
@@ -317,7 +331,8 @@ def save_data_file(g, file_adr):
         "spot_size": g.spot_size,
         "projection": g.projection,
         "alpha": g.alpha,
-        "linewidth": g.linewidth
+        "linewidth": g.linewidth,
+        "depth": g.depth
     }
     data['matplotlib_version'] = matplotlib_version
     result = {"status": True, "message": DATA_SAVE_SUCCESS_MESSAGE}
@@ -358,7 +373,8 @@ def save_config_file(g, file_adr):
         "spot_size": g.spot_size,
         "projection": g.projection,
         "alpha": g.alpha,
-        "linewidth": g.linewidth
+        "linewidth": g.linewidth,
+        "depth": g.depth
     }
     data['matplotlib_version'] = matplotlib_version
     result = {"status": True, "message": DATA_SAVE_SUCCESS_MESSAGE}
@@ -481,6 +497,7 @@ def load_data(g, data):
             g.projection = plot_config.get("projection", DEFAULT_PROJECTION)
             g.alpha = plot_config.get("alpha", DEFAULT_ALPHA)
             g.linewidth = plot_config.get("linewidth", DEFAULT_LINEWIDTH)
+            g.depth = plot_config.get("depth", DEFAULT_DEPTH)
         return
     raise samilaDataError(DATA_TYPE_ERROR)
 
@@ -515,5 +532,6 @@ def load_config(g, config):
             g.projection = plot_config.get("projection", DEFAULT_PROJECTION)
             g.alpha = plot_config.get("alpha", DEFAULT_ALPHA)
             g.linewidth = plot_config.get("linewidth", DEFAULT_LINEWIDTH)
+            g.depth = plot_config.get("depth", DEFAULT_DEPTH)
         return
     raise samilaConfigError(CONFIG_TYPE_ERROR)